GDCDEVOPS-534 add bin scripts to setup.py
Add the indexd bin scripts as "scripts" in setup.py, such that
installing indexd adds those scripts to the PATH.

Ship all such scripts in the Docker image. Make it possible to run
indexd_admin.py and migrate_indexd.py using the container image again.
Note, however, that they are now assumed to be part of the PATH and
are not available at their former location in the /indexd directory.